Enable lazy loading for description and combine with `fetch`
This enables Clerk to cope with moderately sized datasets (gigabytes). Previously
the description could grow unbounded with increasing sizes of
datasets. Now we enable lazy loading for the description and include
the values in it.

Now we also describe all leaf nodes (which were elided before) so the
viewer selection (`:pred` keys) run purely in Clojure.

These changes also enabled downstream simplifications in the sci viewer.