Fix error in describe when sorting sets of maps
In a sorted-set-by if two elements compare to the same score, the one
conjoined later isn't added to the set, e.g. `#{{:a 1} {:b 2}}`.

Prior to this change a `ensure-sorted` called on a set of maps would
return a sorted-set containing just one element as per clojure docs:

We can actually drop `ensure-sorted` in favour of just `sort` because we
don't need to retain set/map types during describe.