Allow setting code cell & result visibility
You can control visibility in Clerk by setting the `:nextjournal.clerk/visibility` which takes a keyword or a set of keywords. Valid values are:

 * `:show` (the default)
 * `:fold` displays the cells collapsed and allows to uncollapse the,
 * `:hide` completely hides the code cells

A declartion on the `ns` form let's all code cells in the notebook inherit the value. On the `ns` form you can also use `:fold-ns` or `:hide-ns` if you'd like an option to only apply to the namespace form. A `ns` form is treated special in that when hidden Clerk will also hide its `nil` result.

The `clerk/hide-result` viewer allows you to hide the result of a code cell.