Markdown Notebooks
Support markdown format for text-heavy notebooks. For now fenced code blocks are evaluated and code listings are not. Each top-level form is displayed with a result so there's no difference in semantics between a Clojure source file and a Markdown file. If this is undesired folks should wrap the cells form in a `do` block for now.