Title & Table of Contents (#56)
This is a first step towards letting Clerk render a table of contents for a given document. We expose :title and :toc attributes as the result of parse for library consumers but do not yet render the table of contents in the dom. The only user facing change derived from this is using the document title in the browser.

Co-authored-by: Martin Kavalar <mk@katercalling.com>