First cut of image viewer (#51)
This adds a default viewer for `java.awt.image.BufferedImage` that can be easily constructed in the JVM using `java.imageio.ImageIO/read`. In this iteration, the viewer will look at the dimensions of the image and pick a suitable layout depending on the size and aspect ratio of the image. Customization and captions will be supported in a follow-up.

Co-authored-by: Martin Kavalar <mk@katercalling.com>