Fix in memory cache not being used for unfreezable results
Prior to this change we would only consider the in-memory results for
results that could also be serialized using nippy. This lead to
unneeded recomputation of results when they contain unserializable
values like Clojure functions. Lifting this restriction should make
recomputation of notebooks fast, regardless of the result value since
the in-memory representation can always be reused.