TOC Redux 🗂 & Dark Mode Toggle 🌗 (#109)
This change brings a table of contents component that is visible by default on desktop. The TOC is opt-in via the `^:nextjournal.clerk/toc` metadata on the ns form (which can be collapsed via `^{:nextjournal.clerk/toc :collapsed}`). In addition to that, dark mode is now also exposed via its own toggle. TOC and dark mode state is persisted in localStorage.

Co-authored-by: Martin Kavalar <mk@katercalling.com>
Co-authored-by: Matt Huebert <mhuebert@gmail.com>