First cut of Clerk Examples (#161)
Add the `nextjournal.clerk/example` macro that evaluates to `nil`, just like `clojure.core/comment` when used outside of Clerk. When used in the context of Clerk it renders the expressions with their resulting values.

Also add `nextjournal.clerk.viewer/mark-preserve-keys` to opt out of transforming maps and their keys.