Use valuehash for deref expressions (#187)
Clerk will now treat `clojure.core/deref` expresssions seperately in the dependency graph and attempt to compute a hash at runtime based on the value of the expression. This lets Clerk see an updated value for these expressions without needing to opt out of Clerk's caching.

In addition, folks can now specify a `:nextjournal.clerk/hash-fn` that makes the hashing user-extensible. This is considered experimental and subject to change.