Cut down on SCI viewer dependencies (#203)
A first step toward reducing the deps of the sci-viewer ns.

- Drops re-frame dep
- Drops devcards and replaces it with a Clerk notebook 

Will work towards dropping the dependency on the nextjournal viewers in a follow-up.

In addition, this improves the inspection of JS objects: prior to this we were falling back to the fallback viewer (pr-str + read + inspect). In particular it now doesn't fail to display nested objects and we can e.g. inspect `js/window`.