Support visibility markers to change default visibility for doc
This adds support for changing the default visibility of code cells
and results within the doc using visibility markers, e.g.

    {:nextjournal.clerk/visibility {:code :fold :result :hide}}

Allows to for example hide the code cells at the beginning of the
document and only show cells below the marker.

The markers themselves will automatically be hidden.

Also assign stable unique ids to the blocks.