First cut of atom synced between JVM ↔ browser (#253)
When defining an atom with `:nextjournal.clerk/sync` metadata on the form, it will now get synced to the browser and interned in the sci env:
```clojure
^:nextjournal.clerk/sync
(defonce my-state
  (atom {:counter 0}))
```

This works in principle but is inefficient because we're currently sending the whole notebook as a result of a recompute from the JVM to the bowser. Making this efficient by sending more granular changes is up next.