Let viewer opt out of var-from-def unwrapping
This fixes an inconsistency in the viewer api: until now we'd unwrap
a `:nextjournal.clerk/var-from-def` when a viewer is applied using an
`fn?` like `clerk/table` but not when given a vieewr map.

We now always unwrap the var unless the viewer opts out with a truthy
`:nextjournal.clerk/var-from-def` key.