Viewer class customization (#294)
This is a first draft for providing custom classes to viewers and the notebook viewer which should allow for most use cases and does not require actually overriding the base styles. Once a `:nextjournal.clerk/css-class` is available on the viewer or in document settings, the available class will be used and no further viewer classes will be assigned.

Also add validation for `:nextjournal.clerk/width` fixing #217.