Drop viewer-js-hash from repo, compute it at runtime (#347)
This drops the viewer-hash-js from the git repo, it was annoying as it always lead to conflicts. Instead we calculate the hash on startup in Clerk dev and when Clerk is used as a git dep. This step is skipped in the jar.

When building a jar we now expect to see a `Skipping coordinate` warning.

Also note that you might need to run `bb build+upload-viewer-resources` before the jar build.

Co-authored-by: Michiel Borkent <michielborkent@gmail.com>