Make `build-graph` recur until all transitive deps are analyzed (#381)
Until now Clerk did not analyze the full transitive dependency graph which could lead to Clerk not detecting a change properly. Analysis is now recursive which means it's taking a bit longer initially. We cache analysis results per file in memory so subsequent analysis should be fast. We will follow up with visualizing the progress of analysis & execution.

Also discovered cases where classes instead of symbols could end up in the dependency graph and introduced normalization to symbols.

This also gets rid of the `->hash must be ifn?` warning which fixes #375.

Co-authored-by: Martin Kavalar <martin@nextjournal.com>
Co-authored-by: Michiel Borkent <michielborkent@gmail.com>