Make following `clerk/doc-url` links usable in interactive mode (#437)
Support following `clerk/doc-url` links in interactive mode. Previously these links would only be functional in the static build. Update the browser url accordingly and support evaluating a given doc by entering it in the browser's address bar.

Also perform an internal refactor to drop the `!error` state atoms in both the webserver and render namespaces and store that state under a key in the `!doc` atom instead.

Change `nextjournal.clerk.render/clerk-eval` to not recompute the currently shown document when using the 1-arity version. Added a second arity that takes an opts map with a `:recompute?` key.

Co-authored-by: Martin Kavalar <martin@nextjournal.com>
Co-authored-by: Philippa Markovics <philippa@markovics.com>