Update Clerk MVN version to the latest in the book
The version of Clerk mentioned in the book right now is updated.

The problem this caused (for me) was with the
`:nextjournal.clerk/visibility` clause, which was a set in the
previous version and is a map now. The documentation suggests using a
map, but the execution fails since it expects a set.

Fixes: #492