Speed up analysis of gitlibs using git sha, resolve protocol methods
This significantly speeds up analysis for gitlibs by using the git sha
as a hash (thus treating them as immutable) instead of handling them
on a per-form level.

Also resolve protocol methods to the defining protocol, which would
previously not be detected.

Lastly drop the location cache which is no longer needed.