Fix preserving expanded-at on clerk/show (#548)
We are resetting a result's expansion state when its `:nextjourna/hash` is changing. The hash is computed starting from the presented value using nippy/freeze. It turns out that adding pagination continuation functions (#421) as metadata on the presented result broke nippy serialization and we'd get a fresh hash (a gensym) on each call to show as a fallback. We can fix this by excluding metadata when nippy freezing.