🦜 Cache remote assets using a ServiceWorker in the browser (#555)
This introduces a ServiceWorker to intercept and cache network requests to remote assets in the browser. It works for Clerk's js bundle, its tailwind css script, fonts and as well as javascript dynamically loaded using d3-require like Clerk's Vega and Plotly viewers.

To use it, you need to open Clerk in the browser when online to populate the cache. Viewers that are dynamically loaded (e.g. Vega or Plotly) need to be used once while offline to be cached. We're considering loading them on worker init in a follow up.

We currently apply it to the following urls:
* https://fonts.bunny.net
* https://cdn.skypack.dev
* https://cdn.tailwindcss.com
* https://storage.clerk.garden
* https://cdn.jsdelivr.net
* https://vega.github.io


Co-authored-by: Martin Kavalar <martin@nextjournal.com>