Simplify handling of Fragments (#598)
Use the recently introduced cell-viewer to simplify the handling of fragments. Fragments are now eagerly unwinded _before_ the presentation of each document block. This also makes the application of the result viewer explicit, it will no longer be lazily matched against a predicate on qualified keys. 

Also fixes the display of defined values in fragments.