Drop ViewerEval type and implement it on top of ViewerFn (#713)
This drops the custom `ViewerEval` type and implements the functionality on top of `ViewerFn` instead. We extend the reading and writing of `ViewerFn` so it can carry an opts map and use that to transmit an `:eval` boolean flag. Also drop the cherry readers in favor of an an `:render-evaluator` opt.

Also switch transmission of unreadable symbols and keywords to a `#clerk/unreadble-edn` reader tag.