Experiment to render viewers to html string (#778)
This is an experiment to find out what's required to support rendering to HTML strings via hiccup, using an alternative viewer stack. If this is feasible, it could be the basis for a much smaller client-side bundle in combination with a morphing lib while moving some of the viewers implementation to custom elements.